Split data retrieval from smart playlists file
